chore: Use integration test
